Refactoring done.

+++ Application/DTOs/OrderDto.cs
@@ -3,7 +3,7 @@ namespace Application.DTOs;
 public class OrderDto
 {
     public int Id { get; set; }
-    public int Sum { get; set; }
+    public decimal Sum { get; set; }
     public DateTime Date { get; set; }
     public int StafferId { get; set; }
     public string StafferFullName { get; set; } = string.Empty;
