+++ Domain/OrderModel.cs
@@ -3,7 +3,7 @@
 public class OrderModel
 {
     public virtual int Id { get; set; }
-    public virtual int Sum { get; set; }
+    public virtual decimal Sum { get; set; }
     public virtual DateTime Date { get; set; }
     public virtual required StafferModel Staffer { get; set; }
     public virtual required CounterAgentModel CounterAgent { get; set; }
